+++ src/Global/GlobalDefinitions.h
@@ -15,7 +15,7 @@ constexpr const u64 hyperblock_bits = 1ULL << 32ULL;
 constexpr const u64 superblock_bits = 1024;
 constexpr const u64 basicblock_bits = 256;
 constexpr const u64 presearch_letters = 12;
-constexpr const uint threads_per_block = 1024;
+constexpr const u64 threads_per_block = 1024;
 
 }  // namespace sbwt_search
