+++ src/Presearcher/Presearcher.cpp
@@ -17,7 +17,7 @@ Presearcher::Presearcher(shared_ptr<GpuSbwtContainer> container_):
 
 auto Presearcher::presearch() -> void {
   constexpr const auto presearch_times
-    = round_up<size_t>(1ULL << (presearch_letters * 2), threads_per_block);
+    = round_up<u64>(1ULL << (presearch_letters * 2), threads_per_block);
   auto blocks_per_grid = presearch_times / threads_per_block;
   auto presearch_left = make_unique<GpuPointer<u64>>(presearch_times);
   auto presearch_right = make_unique<GpuPointer<u64>>(presearch_times);
